+++ api/v1/store_env.go
@@ -16,6 +16,13 @@ const (
 	DatabaseTLSKeyFile    = DatabaseTLSMountPath + "/tls.key"
 )
 
+// Set by the cmd main
+var operatorServiceURL string
+
+func SetOperatorServiceURL(value string) {
+	operatorServiceURL = value
+}
+
 func (s Store) GetDatabaseTLSVolumes() []corev1.Volume {
 	if s.Spec.Database.TLS.SecretName == "" {
 		return nil
@@ -515,7 +522,8 @@ func (s *Store) getFastly() []corev1.EnvVar {
 		s.Spec.ShopConfiguration.Fastly.ServiceRef.Key != "" &&
 		s.Spec.ShopConfiguration.Fastly.TokenRef.Name != "" &&
 		s.Spec.ShopConfiguration.Fastly.TokenRef.Key != "" {
-		envVars = append(envVars,
+		envVars = append(
+			envVars,
 			corev1.EnvVar{
 				Name: "FASTLY_SERVICE_ID",
 				ValueFrom: &corev1.EnvVarSource{
@@ -539,6 +547,18 @@ func (s *Store) getFastly() []corev1.EnvVar {
 				},
 			},
 		)
+
+		// Disable snippet sync from deployment-helper
+		if s.Spec.ShopConfiguration.Fastly.DisableOnSetupMigration {
+			envVars = append(
+				envVars,
+				corev1.EnvVar{
+					Name:  "FASTLY_DISABLE_SNIPPET_UPDATE",
+					Value: "1",
+				},
+			)
+		}
+
 	}
 	return envVars
 }
@@ -647,6 +667,10 @@ func (s *Store) GetEnv() []corev1.EnvVar {
 		{
 			Name:  "DATABASE_PERSISTENT_CONNECTION",
 			Value: "0",
+		},
+		{
+			Name:  "SHOPWARE_OPERATOR_URL",
+			Value: operatorServiceURL,
 		},
 	}